requirements.txt
anthropic
streamlit
pandas
python-dotenv
requests
beautifulsoup4
plotly
